BlastBufferQueue: Update layer size immediately if the buffer can scale
SurfaceView size changes should be reflected immediately instead of
waiting for a new buffer. If the scaling mode is not freeze, the
expected behaviour is for the layer size to be update immediately. If
a buffer is present, it will be scaled to the new size.

If the buffer has scaling mode freeze, then the size changes will
continue to be deferred until we recieve the new buffer of the
new size.

Fixes: 174217687
Test: atest SurfaceViewBufferTests:GeometryTests
Change-Id: I19b6378de63be905db3d9c709cce0f4f26acce18