Set ownerUid/ownerPid for layer with no input info
..instead of callingUid/callingPid. Using the caller instead of the
owner is problematic because it can block touches when it shouldn't.

The example we saw was twitter using an exit animation on their own app,
the window had no input info but the touch was being blocked even
thought it was twitter over twitter, this was because the uid being
assigned was 1000 (the system) instead of twitter's UID.

We have put an exemption for animations that takes into consideration
the alpha, but this is orthogonal, twitter could have any other
animation and should be allowed to let touches pass through to their own
app if they wanted.

Test: Verify windows in exit animations have UID correctly assigned, and
      verify touches are allowed when it's the same app below.
Test: CTS coming
Bug: 173317413
Change-Id: I03ac5dc590456c1b2941e5f351bc32890d1bbca3