Avoid blocking untrusted touches for touch_flag_obscured
Prevent the feature from blocking the touch since we want to verify that
the proper flag is correctly assigned to the motion event.

The test still has value since for some window types (eg. SAW) we might
allow touches via go/untrusted-touches but we still want the
FLAG_WINDOW_IS_OBSCURED mechanism to work.

So, we just configure the surface as TouchOcclusionMode::ALLOW, which
makes go/untrusted-touches bypass it.

Bug: 158002302
Test: atest InputSurfacesTest#touch_flag_obscured
Change-Id: I9fb2846752b8674eec50b777c203a13cdc836144