BlastBufferQueue: Handle queue to window composer queries
When queuing protected buffers, libstagefright queries
ANativeWindow to check if the ANativeWindow sends buffers
directly to SurfaceFlinger. This check works by verifying
the IGBP matches one of the tracked IGBPs in SurfaceFlinger.

This does not apply to blast layers since SurfaceFlinger does
not create IGBPs and there may not be any IGBP used to
submit buffers. To fix this for the blast adapter, we query the
blast buffer queue producer to see if it queues to
SurfaceFlinger.

Bug: b/175904021, b/168917217

Test: atest BlastBufferQueueTest
Test: Play protected content via Google Play on AndroidTV

Change-Id: Id1affce09742c6b438c7499b525615b8bd294b9a