Fix null annotations in AImageDecoder
Bug: 160984428
Test: I27033454b2b8b683db727190b834b970c9732d8e

Make AImageDecoder_delete _Nullable. This matches
AImageDecoderFrameInfo_delete. It also makes sense because it just calls
"delete", and even if it didn't, it does conceptually the same thing, so
it should be safe to pass null.

Fix annotation for AImageDecoder_create*. Although the pointer must not
be null, the pointer it points to can, both as input and as output. This
change doesn't seem to affect my local build system (it doesn't warn if
I pass a pointer to a pointer which points to null without this change),
this is more accurate.

Change-Id: I0bf4e160b21169b81a377378bb65cbfd2ef809b2