AImageDecoder: Hide support for animated HEIF
Bug: 160984428
Bug: 170672785
Test: none

The system currently requires decoding the pixels of the frame of an
HEIF/AVIF animated image before the actual duration of that frame can
be retrieved. Rather than handling this (likely temporary) limitation by
- documenting it
- providing extra APIs
- limiting other formats in the same way
remove documentation specifying that animated HEIF is supported.
Developers are likely to be more interested in the other formats (GIF,
WebP) anyway. If they do attempt to animate an HEIF file, they will
still see the default duration specified for the sequence.

Change-Id: I5b97aeddee1311b694d73ac67b6157032d9ad745