Fix for-loop type.
Technically the iterator provides a std::pair<const KEY, VALUE>&, so
we're constructing a std::pair<KEY, VALUE> from it and then taking a
const reference to that. This creates an unnecessary copy but then
obscures that by taking a reference to it; we should either take a
reference to the iterator's actual type or use the value type of the
temporary we construct directly.

Bug: N/A
Test: compile
Change-Id: I0ed11c1a80a5ac419f74f100573d4a2f042141fd