Increase tolerance for ChildLayerTest, ChildLayerAlpha
Test currently fails on Skia. Increasing the tolerance will let
the test pass while we figure out where is the discrepancy.

Capturing a Skia Debug trace shows that the screen received
pixels of the right color, but the screenshot that the test
sees does not.

Last frame at (0,0):
https://screenshot.googleplex.com/4Z4AJVdzaTzpEKE
Frame 6 at (0,0):
https://screenshot.googleplex.com/AUyKFRzZf68KxPM

Test: set RE backend to "skiaglthreaded"
      adb shell /data/nativetest64/SurfaceFlinger_test/SurfaceFlinger_test --gtest_filter=ChildLayerTest.ChildLayerAlpha
Bug: 175352694
Change-Id: I90370ac12ad7c5f734c1f9b34b837d222c428e7b