SurfaceFlinger: Null guard for setRelativeLayer
Previous versions of this code (up until a change in October) would
simply return false from setRelativeLayer when a null handle was passed.
The current code will have a null SurfaceControl in this situation and
invoking getHandle will crash. It's unclear why a null handle is being
passed to setRelativeLayer but as always SurfaceFlinger should be
resistant to invalid input.

Bug: 178506688
Test: Existing tests pass
Change-Id: Ie6d497b14ec23c848fde3aae4b4940a3ec024f70