Add support for crossfading blur with original input.
If we know that we need to sample the destination for the blur,
then we go ahead and render contents into an offscreen to avoid
an extra render target switch and texture copy.

If the radius is less than kMaxCrossFadeRadius then the final
composition of the blur mixes with the original content to hide
any downscaling artifacts.

Bug: 178503959
Bug: 176903027
Test: notification shade and SilkFX test app
Change-Id: If836461c1c62020447797e9b2d1050156d9c7d0c