Enable offscreen buffers in mskp captures for RenderEngine
This CL also fixes a threading bug in mskp captures. Previously,
if the mskp timer fired in the middle of RE drawing a frame it would
crash because the timer thread would delete the canvas. The mutex
that was introduced will prevent the canvas from being deleted until
after the frame has completed drawing.

Test: capture mskp with background blur
Bug: 176903027
Change-Id: I59442d729b17a7edfd6df7ff9295864d1d67f0d7