Remove token invalid check for BufferSurfaceFrame
Since we already check for token invalid cases while attributing jank
and tracing, we shouldn't perpetuate this to the creation of
SurfaceFrames. Doing so can also result in a case where a buffer is
changed but we don't attribute it to a SurfaceFrame. This change makes
it so that the State's buffer and BufferSurfaceFrame are tied together
and change together.

Bug: 179940619
Test: libsurfaceflinger_unittest
Change-Id: Id420e5924f162d703adc6831edebf4f100db12de