Cleanup Skia RE display and layer colorTransforms
In the general case we don't set the display's colorMatrix on the paint
if we know it is the identity matrix.  We also now abort early if
the layer has no alpha and no colortransform or colorspace conversion
will change that.

The layer's colortransform is now respected regardless of whether or
not color management is enabled for the RE instance

Also includes some small code refactors to avoid computing the same
values in multiple places.

Test: librenderengine_test
Bug: 180712498
Bug: 181028875
Change-Id: I6e236efc5987d00cb464e8798c48f2b3b21635c5